requirements.txt
Django==1.10.3
django-autoslug==1.9.3
Pillow==3.4.2
psycopg2==2.6.2
